+++ core/launcher/api_container_launcher/api_container_launcher.go
@@ -14,18 +14,12 @@ import (
 	"github.com/kurtosis-tech/kurtosis-core/launcher/args"
 	"github.com/kurtosis-tech/stacktrace"
 	"github.com/sirupsen/logrus"
-	"net"
 )
 
 const (
 	// !!!!!!!!!!!!!!!!!! DO NOT MODIFY THIS! IT WILL BE UPDATED AUTOMATICALLY DURING THE RELEASE PROCESS !!!!!!!!!!!!!!!
-	DefaultVersion = "1.45.5"
+	DefaultVersion = "1.47.0"
 	// !!!!!!!!!!!!!!!!!! DO NOT MODIFY THIS! IT WILL BE UPDATED AUTOMATICALLY DURING THE RELEASE PROCESS !!!!!!!!!!!!!!!
-
-	// TODO REMOVE THIS WHEN WE SWITCH TO ENCLAVE DATA VOLUME
-	// The location where the enclave data directory (on the Docker host machine) will be bind-mounted
-	//  on the API container
-	enclaveDataDirpathOnAPIContainer = "/kurtosis-enclave-data"
 
 	enclaveDataVolumeDirpath = "/kurtosis-data"
 
@@ -45,14 +39,9 @@ func (launcher ApiContainerLauncher) LaunchWithDefaultVersion(
 	ctx context.Context,
 	logLevel logrus.Level,
 	enclaveId enclave.EnclaveID,
-	networkId string,
-	subnetMask string,
 	grpcListenPort uint16,
 	grpcProxyListenPort uint16,
-	gatewayIpAddr net.IP,
-	apiContainerIpAddr net.IP,
 	isPartitioningEnabled bool,
-	enclaveDataDirpathOnHostMachine string,
 	metricsUserID string,
 	didUserAcceptSendingMetrics bool,
 ) (
@@ -64,14 +53,9 @@ func (launcher ApiContainerLauncher) LaunchWithDefaultVersion(
 		DefaultVersion,
 		logLevel,
 		enclaveId,
-		networkId,
-		subnetMask,
 		grpcListenPort,
 		grpcProxyListenPort,
-		gatewayIpAddr,
-		apiContainerIpAddr,
 		isPartitioningEnabled,
-		enclaveDataDirpathOnHostMachine,
 		metricsUserID,
 		didUserAcceptSendingMetrics,
 	)
@@ -86,40 +70,22 @@ func (launcher ApiContainerLauncher) LaunchWithCustomVersion(
 	imageVersionTag string,
 	logLevel logrus.Level,
 	enclaveId enclave.EnclaveID,
-	networkId string,
-	subnetMask string,
 	grpcPortNum uint16,
 	grpcProxyPortNum uint16,
-	gatewayIpAddr net.IP,
-	apiContainerIpAddr net.IP,
 	isPartitioningEnabled bool,
-	// TODO REMOVE THIS after we release the switch to enclave data volume
-	enclaveDataDirpathOnHostMachine string,
 	metricsUserID string,
 	didUserAcceptSendingMetrics bool,
 ) (
 	resultApiContainer *api_container.APIContainer,
 	resultErr error,
 ) {
-
-	takenIpAddrStrSet := map[string]bool{
-		gatewayIpAddr.String():      true,
-		apiContainerIpAddr.String(): true,
-	}
 	argsObj, err := args.NewAPIContainerArgs(
 		imageVersionTag,
 		logLevel.String(),
 		grpcPortNum,
 		grpcProxyPortNum,
 		string(enclaveId),
-		networkId,
-		subnetMask,
-		apiContainerIpAddr.String(),
-		takenIpAddrStrSet,
 		isPartitioningEnabled,
-		// TODO REMOVE THIS after we release the switch to enclave data volume
-		enclaveDataDirpathOnAPIContainer,
-		enclaveDataDirpathOnHostMachine,
 		metricsUserID,
 		didUserAcceptSendingMetrics,
 		enclaveDataVolumeDirpath,
@@ -128,7 +94,7 @@ func (launcher ApiContainerLauncher) LaunchWithCustomVersion(
 		return nil, stacktrace.Propagate(err, "An error occurred creating the API container args")
 	}
 
-	envVars, err := args.GetEnvFromArgs(argsObj)
+	envVars, ownIpAddressEnvvar, err := args.GetEnvFromArgs(argsObj)
 	if err != nil {
 		return nil, stacktrace.Propagate(err, "An error occurred generating the API container's environment variables")
 	}
@@ -144,11 +110,10 @@ func (launcher ApiContainerLauncher) LaunchWithCustomVersion(
 		ctx,
 		containerImageAndTag,
 		enclaveId,
-		apiContainerIpAddr,
 		grpcPortNum,
 		grpcProxyPortNum,
-		enclaveDataDirpathOnHostMachine,
 		enclaveDataVolumeDirpath,
+		ownIpAddressEnvvar,
 		envVars,
 	)
 	if err != nil {