+++ core/server/api_container/server/api_container_service.go
@@ -13,6 +13,7 @@ import (
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/module"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/port_spec"
 	kurtosis_backend_service "github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/service"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/user_service_registration"
 	"github.com/kurtosis-tech/kurtosis-core/api/golang/kurtosis_core_rpc_api_bindings"
 	"github.com/kurtosis-tech/kurtosis-core/api/golang/lib/binding_constructors"
 	"github.com/kurtosis-tech/kurtosis-core/server/api_container/server/module_store"
@@ -54,7 +55,7 @@ type ApiContainerService struct {
 
 	enclaveDataDir *enclave_data_directory.EnclaveDataDirectory
 
-	serviceNetwork service_network.ServiceNetwork
+	serviceNetwork *service_network.ServiceNetwork
 
 	moduleStore *module_store.ModuleStore
 
@@ -63,7 +64,7 @@ type ApiContainerService struct {
 
 func NewApiContainerService(
 	enclaveDirectory *enclave_data_directory.EnclaveDataDirectory,
-	serviceNetwork service_network.ServiceNetwork,
+	serviceNetwork *service_network.ServiceNetwork,
 	moduleStore *module_store.ModuleStore,
 	metricsClient client.MetricsClient,
 ) (*ApiContainerService, error) {
@@ -167,23 +168,23 @@ func (apicService ApiContainerService) GetModuleInfo(ctx context.Context, args *
 }
 
 func (apicService ApiContainerService) RegisterService(ctx context.Context, args *kurtosis_core_rpc_api_bindings.RegisterServiceArgs) (*kurtosis_core_rpc_api_bindings.RegisterServiceResponse, error) {
-	serviceId := kurtosis_backend_service.ServiceID(args.ServiceId)
+	serviceId := user_service_registration.ServiceID(args.ServiceId)
 	partitionId := service_network_types.PartitionID(args.PartitionId)
 
-	privateIpAddr, err := apicService.serviceNetwork.RegisterService(serviceId, partitionId)
+	privateIpAddr, err := apicService.serviceNetwork.RegisterService(ctx, serviceId, partitionId)
 	if err != nil {
 		// TODO IP: Leaks internal information about API container
 		return nil, stacktrace.Propagate(err, "An error occurred registering apicService '%v' in the apicService network", serviceId)
 	}
 
 	return &kurtosis_core_rpc_api_bindings.RegisterServiceResponse{
-		PrivateIpAddr:          privateIpAddr.String(),
+		PrivateIpAddr:           privateIpAddr.String(),
 	}, nil
 }
 
 func (apicService ApiContainerService) StartService(ctx context.Context, args *kurtosis_core_rpc_api_bindings.StartServiceArgs) (*kurtosis_core_rpc_api_bindings.StartServiceResponse, error) {
 	logrus.Debugf("Received request to start apicService with the following args: %+v", args)
-	serviceId := kurtosis_backend_service.ServiceID(args.ServiceId)
+	serviceId := user_service_registration.ServiceID(args.ServiceId)
 	privateApiPorts := args.PrivatePorts
 	privateServicePortSpecs := map[string]*port_spec.PortSpec{}
 	for portId, privateApiPort := range privateApiPorts {
@@ -228,14 +229,14 @@ func (apicService ApiContainerService) StartService(ctx context.Context, args *k
 			publicServicePortSpecs,
 		)
 	}
-	logrus.Infof("Started apicService '%v'%v", serviceId, serviceStartLoglineSuffix)
+	logrus.Infof("Started service '%v'%v", serviceId, serviceStartLoglineSuffix)
 
 	return response, nil
 }
 
 func (apicService ApiContainerService) GetServiceInfo(ctx context.Context, args *kurtosis_core_rpc_api_bindings.GetServiceInfoArgs) (*kurtosis_core_rpc_api_bindings.GetServiceInfoResponse, error) {
 	serviceIdStr := args.GetServiceId()
-	serviceId := kurtosis_backend_service.ServiceID(serviceIdStr)
+	serviceId := user_service_registration.ServiceID(serviceIdStr)
 	privateIpAddr, err := apicService.serviceNetwork.GetServiceRegistrationInfo(serviceId)
 	if err != nil {
 		return nil, stacktrace.Propagate(err, "An error occurred getting the registration info for apicService '%v'", serviceIdStr)
@@ -271,7 +272,7 @@ func (apicService ApiContainerService) GetServiceInfo(ctx context.Context, args 
 }
 
 func (apicService ApiContainerService) RemoveService(ctx context.Context, args *kurtosis_core_rpc_api_bindings.RemoveServiceArgs) (*emptypb.Empty, error) {
-	serviceId := kurtosis_backend_service.ServiceID(args.ServiceId)
+	serviceId := user_service_registration.ServiceID(args.ServiceId)
 
 	containerStopTimeoutSeconds := args.ContainerStopTimeoutSeconds
 	containerStopTimeout := time.Duration(containerStopTimeoutSeconds) * time.Second
@@ -285,12 +286,12 @@ func (apicService ApiContainerService) RemoveService(ctx context.Context, args *
 
 func (apicService ApiContainerService) Repartition(ctx context.Context, args *kurtosis_core_rpc_api_bindings.RepartitionArgs) (*emptypb.Empty, error) {
 	// No need to check for dupes here - that happens at the lowest-level call to ServiceNetwork.Repartition (as it should)
-	partitionServices := map[service_network_types.PartitionID]map[kurtosis_backend_service.ServiceID]bool{}
+	partitionServices := map[service_network_types.PartitionID]map[user_service_registration.ServiceID]bool{}
 	for partitionIdStr, servicesInPartition := range args.PartitionServices {
 		partitionId := service_network_types.PartitionID(partitionIdStr)
-		serviceIdSet := map[kurtosis_backend_service.ServiceID]bool{}
+		serviceIdSet := map[user_service_registration.ServiceID]bool{}
 		for serviceIdStr := range servicesInPartition.ServiceIdSet {
-			serviceId := kurtosis_backend_service.ServiceID(serviceIdStr)
+			serviceId := user_service_registration.ServiceID(serviceIdStr)
 			serviceIdSet[serviceId] = true
 		}
 		partitionServices[partitionId] = serviceIdSet
@@ -332,7 +333,7 @@ func (apicService ApiContainerService) Repartition(ctx context.Context, args *ku
 
 func (service ApiContainerService) PauseService(ctx context.Context, args *kurtosis_core_rpc_api_bindings.PauseServiceArgs) (*emptypb.Empty, error) {
 	serviceIdStr := args.ServiceId
-	serviceId := kurtosis_backend_service.ServiceID(serviceIdStr)
+	serviceId := user_service_registration.ServiceID(serviceIdStr)
 	err := service.serviceNetwork.PauseService(ctx, serviceId)
 	if err != nil {
 		return nil, stacktrace.Propagate(err, "Failed to pause service '%v'", serviceId)
@@ -342,7 +343,7 @@ func (service ApiContainerService) PauseService(ctx context.Context, args *kurto
 
 func (service ApiContainerService) UnpauseService(ctx context.Context, args *kurtosis_core_rpc_api_bindings.UnpauseServiceArgs) (*emptypb.Empty, error) {
 	serviceIdStr := args.ServiceId
-	serviceId := kurtosis_backend_service.ServiceID(serviceIdStr)
+	serviceId := user_service_registration.ServiceID(serviceIdStr)
 	err := service.serviceNetwork.UnpauseService(ctx, serviceId)
 	if err != nil {
 		return nil, stacktrace.Propagate(err, "Failed to unpause service '%v'", serviceId)
@@ -352,7 +353,7 @@ func (service ApiContainerService) UnpauseService(ctx context.Context, args *kur
 
 func (apicService ApiContainerService) ExecCommand(ctx context.Context, args *kurtosis_core_rpc_api_bindings.ExecCommandArgs) (*kurtosis_core_rpc_api_bindings.ExecCommandResponse, error) {
 	serviceIdStr := args.ServiceId
-	serviceId := kurtosis_backend_service.ServiceID(serviceIdStr)
+	serviceId := user_service_registration.ServiceID(serviceIdStr)
 	command := args.CommandArgs
 	exitCode, logOutput, err := apicService.serviceNetwork.ExecCommand(ctx, serviceId, command)
 	if err != nil {
@@ -503,7 +504,7 @@ func (apicService ApiContainerService) StoreWebFilesArtifact(ctx context.Context
 
 func (apicService ApiContainerService) StoreFilesArtifactFromService(ctx context.Context, args *kurtosis_core_rpc_api_bindings.StoreFilesArtifactFromServiceArgs) (*kurtosis_core_rpc_api_bindings.StoreFilesArtifactFromServiceResponse, error) {
 	serviceIdStr := args.ServiceId
-	serviceId := kurtosis_backend_service.ServiceID(serviceIdStr)
+	serviceId := user_service_registration.ServiceID(serviceIdStr)
 	srcPath := args.SourcePath
 
 	fileArtifactUUID, err := apicService.serviceNetwork.CopyFromService(ctx, serviceId, srcPath)
@@ -593,7 +594,7 @@ func (apicService ApiContainerService) waitForEndpointAvailability(
 		err  error
 	)
 
-	privateServiceIp, err := apicService.serviceNetwork.GetServiceRegistrationInfo(kurtosis_backend_service.ServiceID(serviceIdStr))
+	privateServiceIp, err := apicService.serviceNetwork.GetServiceRegistrationInfo(user_service_registration.ServiceID(serviceIdStr))
 	if err != nil {
 		return stacktrace.Propagate(err, "An error occurred getting the registration info for apicService '%v'", serviceIdStr)
 	}