+++ core/server/api_container/server/module_store/module_launcher/module_launcher.go
@@ -11,8 +11,6 @@ import (
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/enclave"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/module"
-	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/port_spec"
-	"github.com/kurtosis-tech/free-ip-addr-tracker-lib/lib"
 	"github.com/kurtosis-tech/kurtosis-core/api/golang/kurtosis_core_rpc_api_bindings"
 	"github.com/kurtosis-tech/kurtosis-core/api/golang/module_launch_api"
 	"github.com/kurtosis-tech/kurtosis-core/server/commons/current_time_str_provider"
@@ -27,25 +25,7 @@ const (
 	waitForModuleAvailabilityTimeout = 10 * time.Second
 
 	modulePortNum      = uint16(1111)
-	modulePortProtocol = port_spec.PortProtocol_TCP
-
-	// The filepath where the enclave data directory will be mounted on the module container
-	enclaveDataDirMountFilepathOnContainer = "/kurtosis-enclave-data"
-
-	// This should probably be configurable
-	shouldPullContainerImageBeforeStarting = true
-
-	// Modules don't need to access the Docker engine directly, instead doing that through the API container
-	shouldBindMountDockerSocket = false
-
-	// Indicates that no alias should be set for the module
-	moduleAlias = ""
 )
-
-// These values indicate "don't override the ENTRYPOINT/CMD args" (since modules are configured via envvars)
-var entrypointArgs []string = nil
-var cmdArgs []string = nil
-var volumeMounts map[string]string = nil
 
 type ModuleLauncher struct {
 	// The ID of the enclave that the API container is running inside
@@ -55,19 +35,10 @@ type ModuleLauncher struct {
 
 	// Modules have a connection to the API container, so the launcher must know what socket to pass to modules
 	apiContainerSocketInsideNetwork string
-
-	enclaveDataDirPath string
-	freeIpAddrTracker  *lib.FreeIpAddrTracker
 }
 
-func NewModuleLauncher(enclaveId enclave.EnclaveID, kurtosisBackend backend_interface.KurtosisBackend, apiContainerSocketInsideNetwork string, enclaveDataDirPath string, freeIpAddrTracker *lib.FreeIpAddrTracker) *ModuleLauncher {
-	return &ModuleLauncher{
-		enclaveId:                       enclaveId,
-		kurtosisBackend:                 kurtosisBackend,
-		apiContainerSocketInsideNetwork: apiContainerSocketInsideNetwork,
-		freeIpAddrTracker:               freeIpAddrTracker,
-		enclaveDataDirPath:              enclaveDataDirPath,
-	}
+func NewModuleLauncher(enclaveId enclave.EnclaveID, kurtosisBackend backend_interface.KurtosisBackend, apiContainerSocketInsideNetwork string) *ModuleLauncher {
+	return &ModuleLauncher{enclaveId: enclaveId, kurtosisBackend: kurtosisBackend, apiContainerSocketInsideNetwork: apiContainerSocketInsideNetwork}
 }
 
 func (launcher ModuleLauncher) Launch(
@@ -83,17 +54,11 @@ func (launcher ModuleLauncher) Launch(
 	suffix := current_time_str_provider.GetCurrentTimeStr()
 	moduleGUID := module.ModuleGUID(string(moduleID) + "-" + suffix)
 
-	privateIpAddr, err := launcher.freeIpAddrTracker.GetFreeIpAddr()
-	if err != nil {
-		return nil, nil, stacktrace.Propagate(err, "An error occurred getting a free IP address for new module")
-	}
-
 	args := module_launch_api.NewModuleContainerArgs(
 		string(launcher.enclaveId),
 		modulePortNum,
 		launcher.apiContainerSocketInsideNetwork,
 		serializedParams,
-		enclaveDataDirMountFilepathOnContainer,
 	)
 	envVars, err := module_launch_api.GetEnvFromArgs(args)
 	if err != nil {
@@ -106,9 +71,7 @@ func (launcher ModuleLauncher) Launch(
 		launcher.enclaveId,
 		moduleID,
 		moduleGUID,
-		privateIpAddr,
 		modulePortNum,
-		launcher.enclaveDataDirPath,
 		envVars,
 	)
 
@@ -134,7 +97,7 @@ func (launcher ModuleLauncher) Launch(
 		}
 	}()
 
-	moduleSocket := fmt.Sprintf("%v:%v", privateIpAddr, modulePortNum)
+	moduleSocket := fmt.Sprintf("%v:%v", createdModule.GetPrivateIp(), modulePortNum)
 	conn, err := grpc.Dial(
 		moduleSocket,
 		grpc.WithInsecure(), // TODO SECURITY: Use HTTPS to verify we're connecting to the correct module