+++ core/server/api_container/server/service_network/user_service_launcher/files_artifact_expander/files_artifact_expander.go
@@ -13,21 +13,16 @@ import (
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/files_artifact_expander"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/files_artifact_expansion_volume"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/service"
-	"github.com/kurtosis-tech/free-ip-addr-tracker-lib/lib"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/user_service_registration"
 	"github.com/kurtosis-tech/kurtosis-core/server/commons/current_time_str_provider"
 	"github.com/kurtosis-tech/kurtosis-core/server/commons/enclave_data_directory"
 	"github.com/kurtosis-tech/object-attributes-schema-lib/schema"
 	"github.com/kurtosis-tech/stacktrace"
 	"github.com/sirupsen/logrus"
-	"path"
 	"strings"
 )
 
 const (
-	// Dirpath on the artifact expander container where the enclave data dir (which contains the artifacts)
-	//  will be bind-mounted
-	enclaveDataDirMountpointOnExpanderContainer = "/enclave-data"
-
 	// Dirpath on the artifact expander container where the destination volume containing expanded files will be mounted
 	destVolMntDirpathOnExpander = "/dest"
 
@@ -41,27 +36,22 @@ Class responsible for taking an artifact containing compressed files and creatin
 	into the new volume, and this volume will be mounted on a new service
 */
 type FilesArtifactExpander struct {
-	// Host machine dirpath so the expander can bind-mount it to the artifact expansion containers
-	enclaveDataDirpathOnHostMachine string
-
 	kurtosisBackend backend_interface.KurtosisBackend
 
 	enclaveObjAttrsProvider schema.EnclaveObjectAttributesProvider
 
 	enclaveId enclave.EnclaveID
 
-	freeIpAddrTracker *lib.FreeIpAddrTracker
-
 	filesArtifactStore *enclave_data_directory.FilesArtifactStore
 }
 
-func NewFilesArtifactExpander(enclaveDataDirpathOnHostMachine string, kurtosisBackend backend_interface.KurtosisBackend, enclaveObjAttrsProvider schema.EnclaveObjectAttributesProvider, enclaveId enclave.EnclaveID, freeIpAddrTracker *lib.FreeIpAddrTracker, filesArtifactStore *enclave_data_directory.FilesArtifactStore) *FilesArtifactExpander {
-	return &FilesArtifactExpander{enclaveDataDirpathOnHostMachine: enclaveDataDirpathOnHostMachine, kurtosisBackend: kurtosisBackend, enclaveObjAttrsProvider: enclaveObjAttrsProvider, enclaveId: enclaveId, freeIpAddrTracker: freeIpAddrTracker, filesArtifactStore: filesArtifactStore}
+func NewFilesArtifactExpander(kurtosisBackend backend_interface.KurtosisBackend, enclaveObjAttrsProvider schema.EnclaveObjectAttributesProvider, enclaveId enclave.EnclaveID, filesArtifactStore *enclave_data_directory.FilesArtifactStore) *FilesArtifactExpander {
+	return &FilesArtifactExpander{kurtosisBackend: kurtosisBackend, enclaveObjAttrsProvider: enclaveObjAttrsProvider, enclaveId: enclaveId, filesArtifactStore: filesArtifactStore}
 }
 
 func (expander FilesArtifactExpander) ExpandArtifactsIntoVolumes(
 	ctx context.Context,
-	serviceGUID service.ServiceGUID, // Service GUID for whom the artifacts are being expanded into volumes
+	registrationGuid user_service_registration.UserServiceRegistrationGUID, // Registration GUID of the service for whom the artifacts are being expanded into volumes
 	artifactUuidsToExpand map[service.FilesArtifactID]bool,
 ) (map[service.FilesArtifactID]files_artifact_expansion_volume.FilesArtifactExpansionVolumeName, error) {
 
@@ -74,20 +64,14 @@ func (expander FilesArtifactExpander) ExpandArtifactsIntoVolumes(
 			return nil, stacktrace.Propagate(err, "An error occurred getting the file for files artifact '%v'", filesArtifactId)
 		}
 
-		artifactRelativeFilepath := artifactFile.GetFilepathRelativeToDataDirRoot()
-		artifactFilepathOnExpanderContainer := path.Join(
-			enclaveDataDirMountpointOnExpanderContainer,
-			artifactRelativeFilepath,
-		)
-
 		filesArtifactExpansionVolume, err := expander.kurtosisBackend.CreateFilesArtifactExpansionVolume(
 			ctx,
 			expander.enclaveId,
-			serviceGUID,
+			registrationGuid,
 			filesArtifactId,
 		)
 		if err != nil {
-			return nil, stacktrace.Propagate(err, "An error occurred creating files artifact expansion volume for user service with GUID '%v' and files artifact ID '%v' in enclave with ID '%v'", serviceGUID, filesArtifactId, expander.enclaveId)
+			return nil, stacktrace.Propagate(err, "An error occurred creating files artifact expansion volume for user service with GUID '%v' and files artifact ID '%v' in enclave with ID '%v'", registrationGuid, filesArtifactId, expander.enclaveId)
 		}
 		volumeName := filesArtifactExpansionVolume.GetName()
 		volumesToDestroyIfSomethingFails[volumeName] = true
@@ -102,11 +86,11 @@ func (expander FilesArtifactExpander) ExpandArtifactsIntoVolumes(
 		if err := expander.runFilesArtifactExpander(
 			ctx,
 			filesArtifactId,
-			serviceGUID,
+			registrationGuid,
 			volumeName,
-			artifactFilepathOnExpanderContainer,
+			artifactFile.GetFilepathRelativeToDataDirRoot(),
 		); err != nil {
-			return nil, stacktrace.Propagate(err, "An error occurred running files artifact expander for user service with GUID '%v' and files artifact ID '%v' and files artifact expansion volume '%v' in enclave with ID '%v'",serviceGUID, filesArtifactId, volumeName, expander.enclaveId)
+			return nil, stacktrace.Propagate(err, "An error occurred running files artifact expander for user service with GUID '%v' and files artifact ID '%v' and files artifact expansion volume '%v' in enclave with ID '%v'", registrationGuid, filesArtifactId, volumeName, expander.enclaveId)
 		}
 
 		artifactIdsToVolNames[filesArtifactId] = volumeName
@@ -125,32 +109,18 @@ func (expander FilesArtifactExpander) ExpandArtifactsIntoVolumes(
 func (expander *FilesArtifactExpander) runFilesArtifactExpander(
 	ctx context.Context,
 	filesArtifactId service.FilesArtifactID,
-	serviceGuid service.ServiceGUID,
+	registrationGuid user_service_registration.UserServiceRegistrationGUID,
 	filesArtifactExpansionVolumeName files_artifact_expansion_volume.FilesArtifactExpansionVolumeName,
-	artifactFilepathOnExpanderContainer string,
+	artifactFilepathRelativeToEnclaveDataVolRoot string,
 ) error {
-	// NOTE: This silently (temporarily) uses up one of the user's requested IP addresses with a node
-	//  that's not one of their services! This could get confusing if the user requests exactly a wide enough
-	//  subnet to fit all _their_ services, but we hit the limit because we have these admin containers too
-	//  If this becomes problematic, create a special "admin" network, one per suite execution, for doing thinks like this?
-	// TODO REMOVE THIS ONCE WE FIX THE STATIC IP PROBLEM!!
-	expanderIpAddr, err := expander.freeIpAddrTracker.GetFreeIpAddr()
-	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred getting a free IP for the files artifact expander")
-	}
-	defer expander.freeIpAddrTracker.ReleaseIpAddr(expanderIpAddr)
-
-	guid := newFilesArtifactExpanderGUID(filesArtifactId, serviceGuid)
-
+	guid := newFilesArtifactExpanderGUID(filesArtifactId, registrationGuid)
 	if _, err := expander.kurtosisBackend.RunFilesArtifactExpander(
 		ctx,
 		guid,
 		expander.enclaveId,
 		filesArtifactExpansionVolumeName,
-		expander.enclaveDataDirpathOnHostMachine,
 		destVolMntDirpathOnExpander,
-		artifactFilepathOnExpanderContainer,
-		expanderIpAddr,
+		artifactFilepathRelativeToEnclaveDataVolRoot,
 	); err != nil {
 		return stacktrace.Propagate(err, "An error occurred running files artifact expander with GUID '%v' for files artifact expansion volume '%v' in enclave with ID '%v'", guid, filesArtifactExpansionVolumeName, expander.enclaveId)
 	}
@@ -192,11 +162,11 @@ func (expander *FilesArtifactExpander) destroyFilesArtifactExpansionVolumes(ctx 
 	}
 }
 
-func newFilesArtifactExpanderGUID(filesArtifactId service.FilesArtifactID, userServiceGuid service.ServiceGUID) files_artifact_expander.FilesArtifactExpanderGUID {
-	userServiceGuidStr := string(userServiceGuid)
+func newFilesArtifactExpanderGUID(filesArtifactId service.FilesArtifactID, serviceRegistrationGuid user_service_registration.UserServiceRegistrationGUID) files_artifact_expander.FilesArtifactExpanderGUID {
+	serviceRegistrationGuidStr := string(serviceRegistrationGuid)
 	filesArtifactIdStr := string(filesArtifactId)
 	suffix := current_time_str_provider.GetCurrentTimeStr()
-	guidStr := strings.Join([]string{userServiceGuidStr, filesArtifactIdStr, suffix}, guidElementSeparator)
+	guidStr := strings.Join([]string{serviceRegistrationGuidStr, filesArtifactIdStr, suffix}, guidElementSeparator)
 	guid := files_artifact_expander.FilesArtifactExpanderGUID(guidStr)
 	return guid
 }